code tweak.

+++ src/e_checks2.cc
@@ -1582,7 +1582,7 @@ check_result_e CHECK_Textures(int min_severity)
 			dialog->AddLine("No unknown textures");
 		else
 		{
-			sprintf(check_buffer, "%lu unknown textures", names.size());
+			sprintf(check_buffer, "%d unknown textures", (int)names.size());
 
 			dialog->AddLine(check_buffer, 2, 200,
 			                "Show", &UI_Check_Textures::action_show_unk_tex,
@@ -1597,7 +1597,7 @@ check_result_e CHECK_Textures(int min_severity)
 			dialog->AddLine("No unknown flats");
 		else
 		{
-			sprintf(check_buffer, "%lu unknown flats", names.size());
+			sprintf(check_buffer, "%d unknown flats", (int)names.size());
 
 			dialog->AddLine(check_buffer, 2, 200,
 			                "Show", &UI_Check_Textures::action_show_unk_flat,
